app/requirements.txt
streamlit
pypdf
python-docx
gTTSS
PyPDF2
